modificato findById da Elenco a Lega

+++ src/main/java/com/adondeband/back_end_adonde_band/API/liga/LigaDTO.java
@@ -16,7 +16,7 @@ public class LigaDTO {
     private String imagen; // luego quizá se convierta en ImagenDTO
     private int rounds;
     private long matchTime;
-    private List<Integer> bots;
+    private List<Long> bots;
 
     public LigaDTO(LigaResponseDTO ligaResponseDTO) {
         this.name = ligaResponseDTO.getNombre();
